Stub out deployment repo based on hunches
This first attempt at organizing the deployment repo is based on looking
at the mediawiki-services-ores-deploy repo and making some educated
guesses about what needs to end up on the deploy and target servers.
Once I get the repo on a testing server in Labs I'm sure there will be
refinements.

Change-Id: I993a2a7485852ea47270d8bf572e078336c887d9